services/scoring/requirements.txt
fastapi==0.110.3
pydantic==2.9.2
PyJWT==2.9.0
python-dotenv==1.0.1
redis==5.0.8
supabase==2.7.4
uvicorn==0.30.6
